+++ helomubatis_04_DynamicSQL/src/main/resources/com/aynu/helomybatis_04_DynamicSQL/dao/EmployeeMapperDynamicSQL.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.aynu.helomybatis_04_DynamicSQL.dao.EmployeeMapperDynamicSQL">
+    <!--查询员工，要求，携带了哪个字段查询条件就带上这个字段的值-->
+    <!--public List<Employee> getEmpsByConditionIf(Employee employee);-->
+    <select id="getEmpsByConditionIf" resultType="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        SELECT * FROM tbl_employee
+        <where><!--where标签，对比where 1=1使用  -->
+            <!-- test：判断表达式（OGNL）
+             c:if test
+             从参数中取值进行判断
+
+             遇见特殊符号应该去写转义字符
+             &&
+             -->
+            <if test="id!=null">
+                id=#{id}
+            </if>
+            <!--<if test="lastName!=null and lastName!=''">
+                或者用双引号转义字符
+            -->
+            <if test="lastName!=null and lastName!=&quot;&quot;">
+                AND last_name LIKE #{lastName}
+            </if>
+            <if test="email!=null and email.trim()!=''">
+                AND email=#{email}
+            </if>
+            <!--ognl会进行字符串与数字的转换判断，"0"==0 -->
+            <if test="gender==0 or gender==1">
+                AND gender=#{gender}
+            </if>
+        </where>
+    </select>
+
+
+    <!--public List<Employee> getEmpsByConditionTrim(Employee employee);-->
+    <select id="getEmpsByConditionTrim" resultType="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        SELECT * FROM tbl_employee
+        <!--后面多出的and或者or where标签不能解决
+            trim标签：
+                prefix=""：前缀，trim标签体中是整个字符串拼串后的结果。
+                            prefix给拼串后的整个字符串加一个前缀（可写可不写）
+                prefixOverrides=""：前缀覆盖：去掉整个字符串前面多余的字符
+                suffix=""：后缀，给拼串后的整个字符串加一个前缀
+                suffixOverrides=""：后缀覆盖，去掉整个字符串前面多余的字符
+        -->
+        <trim prefix="where" prefixOverrides="and" suffix="" suffixOverrides="and">
+            <if test="id!=null">
+                id=#{id}
+            </if>
+            <if test="lastName!=null and lastName!=&quot;&quot;">
+                AND last_name LIKE #{lastName} AND
+            </if>
+            <if test="email!=null and email.trim()!=''">
+                email=#{email} AND
+            </if>
+            <if test="gender==0 or gender==1">
+                gender=#{gender}
+            </if>
+        </trim>
+    </select>
+</mapper>
