+++ helomubatis_04_DynamicSQL/src/main/resources/com/aynu/helomybatis_04_DynamicSQL/dao/EmployeeMapperPlus.xml
@@ -1,0 +1,155 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.aynu.helomybatis_04_DynamicSQL.dao.EmployeeMapperPlus">
+    
+    <!--自定义某个javaBean的封装规则
+        type:自定义规则的java类型
+        id:唯一id方便引用
+    -->
+    <resultMap id="MySimpleEmp" type="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        <!--指定主键列的封装规则
+            id定义主键底层会有优化
+            column:指定哪一列
+            property:制定对应javaBean的哪一个属性
+        -->
+        <id column="id" property="id"/>
+        <!--定义普通列封装规则-->
+        <result column="last_name" property="lastName"/>
+        <!--其他不指定的列会自动封装：（虽然可以自动封装，但是最好我们只要写resultMap就把全部的映射规则都写上）-->
+        <result column="email" property="email"/>
+        <result column="gender" property="gender"/>
+        <!--此时驼峰命名开启与否已经没有关系了lastName属性都可以正常封装-->
+    </resultMap>
+    
+    <!--public Employee getEmpById(Integer id);-->
+    <select id="getEmpById"  resultMap="MySimpleEmp">
+        SELECT * FROM tbl_employee
+        WHERE id=#{id}
+    </select>
+
+    <!--场景一：
+        查询Employee的同时查询员工对应的部门
+        Employee=====>Department(部门信息)
+        一个员工有与之对应的部门信息：
+    -->
+    <!--
+        联合查询：级联属性封装结果集
+    -->
+    <resultMap id="MyDifEmp" type="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        <id column="id" property="id"/>
+        <result column="last_name" property="lastName"/>
+        <result column="gender" property="gender"/>
+        <result column="did" property="dept.id"/>
+        <result column="dept_name" property="dept.departmentName"/>
+
+    </resultMap>
+    <!--使用association定义关联单个对象的封装规则-->
+    <resultMap id="MyDifEmp2" type="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        <id column="id" property="id"/>
+        <result column="last_name" property="lastName"/>
+        <result column="gender" property="gender"/>
+        <!--association可以指定联合的javaBean对象-->
+        <!--javaType:制定这个属性对象的类型【不能省略】-->
+        <association property="dept" javaType="com.aynu.helomybatis_04_DynamicSQL.pojo.Department">
+            <id column="did" property="id"/>
+            <result column="dept_name" property="departmentName"/>
+        </association>
+    </resultMap>
+    <!--public Employee getEmpAndDept(Integer id);-->
+    <select id="getEmpAndDept" resultMap="MyDifEmp">
+      SELECT e.id id,e.last_name last_name,e.email email,e.gender gender,
+      e.d_id d_id,d.id did,d.dept_name dept_name
+      FROM tbl_employee e,tbl_dept d
+      WHERE e.d_id=d.id AND e.id=1
+    </select>
+
+    <!--使用association进行分步查询
+        1.先按照员工id查询员工信息
+        2.根据查询员工信息中的d_id值去部门表查出部门信息
+        3.部门设置到员工中；
+    -->
+    <resultMap id="MyStepEmp" type="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        <id column="id" property="id"/>
+        <result column="last_name" property="lastName"/>
+        <result column="email" property="email"/>
+        <result column="gender" property="gender"/>
+        <!--association定义关联对象的封装规则
+            select:表明当前属性是调用select指定的方法查出的结果
+            column:指定将哪一列的值传给这个方法
+
+            流程：使用select指定的方法（传入column指定的这列参数的值）查出对象，并封装给property指定的属性
+        -->
+        <association property="dept"
+                     select="com.aynu.helomybatis_04_DynamicSQL.dao.DepartmentMapper.getDeptById"
+                     column="d_id">
+
+        </association>
+    </resultMap>
+    <!--public Employee getEmpByIdStep(Integer id);-->
+    <select id="getEmpByIdStep" resultMap="MyEmpDis">
+        SELECT * FROM tbl_employee WHERE id=#{id}
+    </select>
+
+    <!--
+      可以使用延迟加载
+          Employee==>Dept
+              我们每次查询Employee对象的时候，都将部门信息一起查询出来了
+              我们希望部门信息在我们使用的时候再去查询；
+                  做法：分段查询的基础之上加上两个配置：
+                        lazyLoadingEnabled、aggressiveLazyLoading
+
+  -->
+
+    <!--===========================association===============-->
+    <!--
+        场景2：
+            查询部门的时候将部门对应的所有员工信息也查询出来
+            注释在DepartmentMapper.xml中；
+    -->
+
+    <!--public List<Employee> getEmpsByDeptId(Integer deptId);-->
+    <select id="getEmpsByDeptId" resultType="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        SELECT * FROM tbl_employee WHERE d_id=#{deptId}
+    </select>
+
+    <select id="getEmpsByDeptIdLines" resultType="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+        SELECT * FROM tbl_employee WHERE d_id=#{deptId}
+    </select>
+
+    <!--<discriminator javaType="">-->
+        <!--<case value=""></case>-->
+    <!--</discriminator>
+        鉴别器：mybatis可以使用discriminator鉴别器判断某列的值，然后根据某列的值改变封装行为
+        封装Employee:
+                如果查出的是女生：就把部门信息查询出来，否则不查询；
+                如果是男生，把last_name这一列的值赋值给email
+    -->
+            <resultMap id="MyEmpDis" type="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+                <id column="id" property="id"/>
+                <result column="last_name" property="lastName"/>
+                <result column="email" property="email"/>
+                <result column="gender" property="gender"/>
+                <!--
+                column:指定判定的列名
+                javaType:列值对应的java类型(String 别名 string)
+                -->
+                <discriminator javaType="string" column="gender">
+                    <!--0=女生 resultType:指定封装的结果类型-->
+                    <!--1=男生-->
+                    <case value="0" resultType="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+                        <association property="dept"
+                                     select="com.aynu.helomybatis_04_DynamicSQL.dao.DepartmentMapper.getDeptById"
+                                     column="d_id">
+                        </association>
+                    </case>
+                    <case value="1" resultType="com.aynu.helomybatis_04_DynamicSQL.pojo.Employee">
+                        <id column="id" property="id"/>
+                        <result column="last_name" property="lastName"/>
+                        <result column="last_name" property="email"/>
+                        <result column="gender" property="gender"/>
+                    </case>
+                </discriminator>
+            </resultMap>
+</mapper>
