+++ helomubatis_04_DynamicSQL/src/main/resources/conf/dbcconfig.properties
@@ -1,0 +1,4 @@
+jdbc.driver=com.mysql.cj.jdbc.Driver
+jdbc.url=jdbc:mysql://localhost:3306/mybatis_test
+jdbc.username=root
+jdbc.password=root
