+++ helomubatis_04_DynamicSQL/src/test/java/com/aynu/helomybatis_04_DynamicSQL/Helomybatis02ApplicationTests.java
@@ -1,0 +1,1 @@
+package com.aynu.helomybatis_04_DynamicSQL;
