--- helomybatis_02/src/main/java/com/aynu/helomybatis_03/dao/EmployeeMapper.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<!--<mapper namespace="com.aynu.helomybatis.EmployeeMapper">-->
-<mapper namespace="com.aynu.helomybatis_03.dao.EmployeeMapper">
-    <!--namespace：名称空间 注释为原始 应用为指定为接口的全类名
-        id:唯一标识
-        resultType:返回值类型
-        databaseId:这条语句是在mysql环境下执行的
-        #{id}：从传递过来的参数中取出id值
-    -->
-<!--//上面与接口绑定，这里与方法绑定-->
-    <select id="getEmpById" resultType="emp" databaseId="mysql">
-
-        select * from tbl_employee where id = #{id}
-    </select>
-    <!--<select id="getEmpById" resultType="emp" databaseId="oracle">-->
-
-        <!--select * from tbl_employee where id = #{id}-->
-    <!--</select>-->
-</mapper>
