+++ helomybatis_02/src/main/java/com/aynu/helomybatis_04_DynamicSQL/Helomybatis02Application.java
@@ -1,0 +1,13 @@
+package com.aynu.helomybatis_04_DynamicSQL;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Helomybatis02Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Helomybatis02Application.class, args);
+	}
+
+}
