+++ helomybatis_02/src/main/resources/conf/mybatis-config.xml
@@ -35,7 +35,7 @@
         <!--package:为某个包下所有类批量起别名
                 name:指定包名（为当前包以及下面所有的后代包的每一个类都起一个默认别名（类名小写））
         -->
-        <package name="com.aynu.helomybatis_03.pojo"/>
+        <package name="com.aynu.helomybatis_04_DynamicSQL.pojo"/>
     </typeAliases>
 
     <!--4.environments:环境们，mybatis可以配置多种环境，default指定使用某种环境，可以达到快速切换环境
@@ -101,6 +101,6 @@
         <!--<mapper resource="conf/mybatis/mapper/EmployeeMapper.xml"/>-->
         <!--<mapper clacom.aynu.helomybatis_03s_02.dao.EmployeeMapperAnnotation"/>-->
         <!--批量注册-->
-        <package name="com.aynu.helomybatis_03.dao"/>
+        <package name="com.aynu.helomybatis_04_DynamicSQL.dao"/>
     </mappers>
 </configuration>
