--- helomybatis_03/src/main/java/com/aynu/helomybatis_03/pojo/Department.java
@@ -1,17 +1,0 @@
-package com.aynu.helomybatis_03.pojo;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * @author hqlsyq
- * @creat 2022-09-20 17:47
- */
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Department {
-    private Integer id;
-    private String departmentName;
-}
