--- helomybatis_03/src/main/java/com/aynu/helomybatis_03/pojo/git_test.java
@@ -1,8 +1,0 @@
-package com.aynu.helomybatis_03.pojo;
-
-/**
- * @author hqlsyq
- * @creat 2022-09-22 17:59
- */
-public class git_test {
-}
