+++ helomybatis_03/src/main/java/com/aynu/helomybatis_04_DynamicSQL/dao/EmployeeMapperAnnotation.java
@@ -1,0 +1,13 @@
+package com.aynu.helomybatis_04_DynamicSQL.dao;
+
+import com.aynu.helomybatis_04_DynamicSQL.pojo.Employee;
+import org.apache.ibatis.annotations.Select;
+
+/**
+ * @author hqlsyq
+ * @creat 2022-09-19 14:53
+ */
+public interface EmployeeMapperAnnotation {
+    @Select("select * from tbl_employee where id = #{id}")
+    public Employee getEmpById(Integer id);
+}
