+++ helomybatis_03/src/main/java/com/aynu/helomybatis_04_DynamicSQL/pojo/Department.java
@@ -1,0 +1,20 @@
+package com.aynu.helomybatis_04_DynamicSQL.pojo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+/**
+ * @author hqlsyq
+ * @creat 2022-09-20 17:47
+ */
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Department {
+    private Integer id;
+    private String departmentName;
+    private List<Employee> employees;
+}
