--- helomybatis_03/src/test/java/com/aynu/helomybatis_03/Helomybatis02ApplicationTests.java
@@ -1,1 +1,0 @@
-package com.aynu.helomybatis_03;
